README.md
This utility will read in an Activeworlds propdump file (version 5) and create a text file containing a list of all unique models in that dump. The model list text file will be put in the same directory as the propdump. I may at some point in the future add support for v3 and v4 propdumps.